refactor: Update PostController, PostControllerTest
PostDto 추가
PostController 수정
PostService 수정
PostControllerTest 추가 및 수정

+++ src/main/java/com/majorrunner/majorserver/post/PostController.java
@@ -1,11 +1,7 @@
 package com.majorrunner.majorserver.post;
 
-import com.majorrunner.majorserver.Like.Like;
-import com.majorrunner.majorserver.category.Category;
-import com.majorrunner.majorserver.comment.Comment;
-import com.majorrunner.majorserver.user.User;
-import lombok.Data;
 import lombok.RequiredArgsConstructor;
+import org.modelmapper.ModelMapper;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.web.PagedResourcesAssembler;
@@ -17,8 +13,6 @@ import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
 import java.net.URI;
-import java.time.LocalDateTime;
-import java.util.List;
 
 import static org.springframework.hateoas.mvc.ControllerLinkBuilder.linkTo;
 
@@ -27,12 +21,16 @@ import static org.springframework.hateoas.mvc.ControllerLinkBuilder.linkTo;
 @RequestMapping(value = "/api/posts", produces = MediaTypes.HAL_JSON_UTF8_VALUE)
 public class PostController {
 
+    private final ModelMapper modelMapper;
     private final PostRepository postRepository;
     private final PostService postService;
 
     @PostMapping
-    public ResponseEntity createPost(@RequestBody @Valid Post post) {
-        Long postId = postService.add(post);
+    public ResponseEntity createPost(@RequestBody @Valid PostDto postDto) {
+
+        Post post = modelMapper.map(postDto, Post.class);
+
+        postService.add(post);
 
         ControllerLinkBuilder selfLinkBuilder = linkTo(PostController.class).slash(post.getId());
         URI createdUri = selfLinkBuilder.toUri();
@@ -40,20 +38,17 @@ public class PostController {
         return ResponseEntity.created(createdUri).body(post);
     }
 
-    @GetMapping
-    public ResponseEntity queryPosts(Pageable pageable, PagedResourcesAssembler<Post> assbler) {
-
-        Page<Post> page = postRepository.findAll(pageable);
-        PagedResources<PostResource> pagedResources = assbler.toResource(page, e -> new PostResource(e));
-
-        return ResponseEntity.ok().body(pagedResources);
-    }
-
     @GetMapping("/{id}")
     public ResponseEntity getPost(@PathVariable Long id) {
 
         Post post = postService.findOne(id);
-        PostDto postDto = new PostDto(post);
+
+        if (post == null) {
+            return ResponseEntity.notFound().build();
+        }
+
+        PostDto postDto = modelMapper.map(post, PostDto.class);
+
         ControllerLinkBuilder selfLinkBuilder = linkTo(PostController.class).slash(post.getId());
         URI createdUri = selfLinkBuilder.toUri();
 
@@ -61,48 +56,34 @@ public class PostController {
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity updatePost(@PathVariable Long id, Post post) {
+    public ResponseEntity updatePost(@PathVariable Long id, @RequestBody @Valid PostDto postDto) {
+        
+        Post post = postService.findOne(id);
 
-        Post updatedPost = postService.update(post);
+        if (post == null) {
+            return ResponseEntity.notFound().build();
+        }
+
+        postService.update(postDto, post);
 
         ControllerLinkBuilder selfLinkBuilder = linkTo(PostController.class).slash(post.getId());
         URI createdUri = selfLinkBuilder.toUri();
 
-        return ResponseEntity.created(createdUri).body(updatedPost);
+        return ResponseEntity.created(createdUri).build();
     }
 
     @DeleteMapping("/{id}")
     public ResponseEntity deletePost(@PathVariable Long id) {
 
         Post post = postService.findOne(id);
+
+        if (post == null) {
+            return ResponseEntity.notFound().build();
+        }
+
         postService.delete(post);
 
         return ResponseEntity.noContent().build();
     }
 
-    @Data
-    static class PostDto {
-
-        private Long postId;
-        private Article article;
-        private int views;
-        private User user;
-        private Category category;
-        private List<Comment> comments;
-        private List<Like> likes;
-        private LocalDateTime createdAt;
-        private LocalDateTime updatedAt;
-
-        public PostDto(Post post) {
-            postId = post.getId();
-            article = post.getArticle();
-            views = post.getViews();
-            user = post.getUser();
-            category = post.getCategory();
-            comments = post.getComments();
-            likes = post.getLikes();
-            createdAt = post.getCreatedAt();
-            updatedAt = post.getUpdatedAt();
-        }
-    }
 }
